refactor the receiver
